requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
python-docx>=1.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
streamlit>=1.28.0
trino>=0.326.0
